framework_lib: Rename uefi to fwk_uefi
Avoid clash with uefi crate
