+++ framework_lib/src/commandline/mod.rs
@@ -58,7 +58,7 @@ use crate::touchpad::print_touchpad_fw_ver;
 #[cfg(feature = "hidapi")]
 use crate::touchscreen;
 #[cfg(feature = "uefi")]
-use crate::uefi::enable_page_break;
+use crate::fw_uefi::enable_page_break;
 use crate::util::{self, Config, Platform, PlatformFamily};
 #[cfg(feature = "hidapi")]
 use hidapi::HidApi;
@@ -639,7 +639,7 @@ fn print_esrt() {
 
 fn flash_ec(ec: &CrosEc, ec_bin_path: &str, flash_type: EcFlashType, dry_run: bool) {
     #[cfg(feature = "uefi")]
-    let data = crate::uefi::fs::shell_read_file(ec_bin_path);
+    let data = crate::fw_uefi::fs::shell_read_file(ec_bin_path);
     #[cfg(not(feature = "uefi"))]
     let data: Option<Vec<u8>> = {
         match fs::read(ec_bin_path) {
@@ -674,7 +674,7 @@ fn dump_ec_flash(ec: &CrosEc, dump_path: &str) {
     }
     #[cfg(feature = "uefi")]
     {
-        let ret = crate::uefi::fs::shell_write_file(dump_path, &flash_bin);
+        let ret = crate::fw_uefi::fs::shell_write_file(dump_path, &flash_bin);
         if ret.is_err() {
             println!("Failed to dump EC FW image.");
         }
@@ -691,7 +691,7 @@ fn dump_dgpu_eeprom(ec: &CrosEc, dump_path: &str) {
     }
     #[cfg(feature = "uefi")]
     {
-        let ret = crate::uefi::fs::shell_write_file(dump_path, &flash_bin);
+        let ret = crate::fw_uefi::fs::shell_write_file(dump_path, &flash_bin);
         if ret.is_err() {
             println!("Failed to dump EC FW image.");
         }
@@ -1117,7 +1117,7 @@ pub fn run_with_args(args: &Cli, _allupdate: bool) -> i32 {
     //    raw_command(&args[1..]);
     } else if let Some(pd_bin_path) = &args.pd_bin {
         #[cfg(feature = "uefi")]
-        let data: Option<Vec<u8>> = crate::uefi::fs::shell_read_file(pd_bin_path);
+        let data: Option<Vec<u8>> = crate::fw_uefi::fs::shell_read_file(pd_bin_path);
         #[cfg(not(feature = "uefi"))]
         let data = match fs::read(pd_bin_path) {
             Ok(data) => Some(data),
@@ -1136,7 +1136,7 @@ pub fn run_with_args(args: &Cli, _allupdate: bool) -> i32 {
         }
     } else if let Some(ec_bin_path) = &args.ec_bin {
         #[cfg(feature = "uefi")]
-        let data: Option<Vec<u8>> = crate::uefi::fs::shell_read_file(ec_bin_path);
+        let data: Option<Vec<u8>> = crate::fw_uefi::fs::shell_read_file(ec_bin_path);
         #[cfg(not(feature = "uefi"))]
         let data = match fs::read(ec_bin_path) {
             Ok(data) => Some(data),
@@ -1155,7 +1155,7 @@ pub fn run_with_args(args: &Cli, _allupdate: bool) -> i32 {
         }
     } else if let Some(capsule_path) = &args.capsule {
         #[cfg(feature = "uefi")]
-        let data: Option<Vec<u8>> = crate::uefi::fs::shell_read_file(capsule_path);
+        let data: Option<Vec<u8>> = crate::fw_uefi::fs::shell_read_file(capsule_path);
         #[cfg(not(feature = "uefi"))]
         let data = match fs::read(capsule_path) {
             Ok(data) => Some(data),
@@ -1184,7 +1184,7 @@ pub fn run_with_args(args: &Cli, _allupdate: bool) -> i32 {
         }
     } else if let Some(capsule_path) = &args.h2o_capsule {
         #[cfg(feature = "uefi")]
-        let data = crate::uefi::fs::shell_read_file(capsule_path);
+        let data = crate::fw_uefi::fs::shell_read_file(capsule_path);
         #[cfg(not(feature = "uefi"))]
         let data = match fs::read(capsule_path) {
             Ok(data) => Some(data),
@@ -1237,7 +1237,7 @@ pub fn run_with_args(args: &Cli, _allupdate: bool) -> i32 {
     } else if let Some(hash_file) = &args.hash {
         println!("Hashing file: {}", hash_file);
         #[cfg(feature = "uefi")]
-        let data = crate::uefi::fs::shell_read_file(hash_file);
+        let data = crate::fw_uefi::fs::shell_read_file(hash_file);
         #[cfg(not(feature = "uefi"))]
         let data = match fs::read(hash_file) {
             Ok(data) => Some(data),
@@ -1266,7 +1266,7 @@ pub fn run_with_args(args: &Cli, _allupdate: bool) -> i32 {
             Some(PlatformFamily::Framework16) | None
         ) {
             #[cfg(feature = "uefi")]
-            let data: Option<Vec<u8>> = crate::uefi::fs::shell_read_file(gpu_descriptor_file);
+            let data: Option<Vec<u8>> = crate::fw_uefi::fs::shell_read_file(gpu_descriptor_file);
             #[cfg(not(feature = "uefi"))]
             let data = match fs::read(gpu_descriptor_file) {
                 Ok(data) => Some(data),
