+++ framework_lib/src/lib.rs
@@ -42,7 +42,7 @@ mod os_specific;
 pub mod power;
 pub mod smbios;
 #[cfg(feature = "uefi")]
-pub mod uefi;
+pub mod fw_uefi;
 mod util;
 
 pub mod built_info {
